requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
dj-database-url==2.3.0
django-cors-headers==4.6.0
django-filter==1.1.0
djangorestframework==3.16.1
djangorestframework-filters==0.11.1
djangorestframework_simplejwt==5.5.1
google-auth==2.47.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
